Change template rendering implementation
We need to gear up for templates (e.g. pages) being extended from something else, so the actual rendering needs to start from the template itself.

It is a safe assumption that all templates will be extended as it would be somewhat madness to build up the general page structure for a large site without doing so.